refactor(openomni): split worker spawn payload parsing
