+++ packages/openomni/src/dispatch/handlers/worker.ts
@@ -1,5 +1,5 @@
 import { Execution, type Dispatch, type Model } from "@openomni/protocol";
-import { Session, WorkItemStore } from "@openomni/session";
+import { WorkItemStore } from "@openomni/session";
 import { z } from "zod";
 import type { CoordinatorLike } from "../../ingress/coordinator-like.js";
 import type { ConnectorEndpointDriverOwner } from "../owners.js";
@@ -15,6 +15,7 @@ import {
   reflectCoordinatorResult,
   type WorkerCompletionOptions,
 } from "./worker-completion.js";
+import { buildWorkerSpawnRequest, parseWorkerSpawnPayload } from "./worker-spawn-payload.js";
 import { createWorkerSpawnWorkItem } from "./worker-work-item.js";
 import { extractText } from "./shared.js";
 
@@ -24,29 +25,7 @@ export interface WorkerDispatchHandlerOptions extends WorkerCompletionOptions {
   readonly defaultModel?: Model.Ref;
 }
 
-const AcceptanceCriterion = z.string().trim().min(1);
 const INTERNAL_EXECUTOR_KIND = "internal_chat_agent";
-const WORKER_SPAWN_TEXT_OR_PROMPT_MESSAGE = "worker.spawn requires text or prompt";
-const WorkerSpawnPayload = z
-  .object({
-    text: z.string().trim().min(1).optional(),
-    prompt: z.string().trim().min(1).optional(),
-    acceptanceCriteria: z.array(AcceptanceCriterion).min(1),
-    constraints: z.array(z.string().trim().min(1)).optional(),
-  })
-  .strict()
-  .refine((payload) => payload.text !== undefined || payload.prompt !== undefined, {
-    message: WORKER_SPAWN_TEXT_OR_PROMPT_MESSAGE,
-    path: ["text"],
-  })
-  .transform((payload) => ({
-    prompt: payload.text ?? payload.prompt ?? "",
-    acceptanceCriteria: payload.acceptanceCriteria,
-    ...(payload.constraints ? { constraints: payload.constraints } : {}),
-  }));
-
-type WorkerSpawnPayloadInput = z.input<typeof WorkerSpawnPayload>;
-type ParsedWorkerSpawnPayload = z.infer<typeof WorkerSpawnPayload>;
 const WorkerCompletePayload = z
   .object({
     workItemHash: z.string().min(1).optional(),
@@ -62,43 +41,6 @@ function requireCoordinator(coordinator: CoordinatorLike | undefined): Coordinat
 
 function targetRunId(command: Dispatch.Command): string | undefined {
   return command.target.runId ?? command.target.id;
-}
-
-function resolveSessionId(command: Dispatch.Command, model: Model.Ref): string {
-  if (command.target.sessionId) return command.target.sessionId;
-  const title = `Dispatch worker ${command.action}`;
-  const modelInfo = { providerID: model.provider, modelID: model.id };
-  const session = command.target.parentSessionId
-    ? Session.createChild({
-        parentSessionId: command.target.parentSessionId,
-        title,
-        model: modelInfo,
-      })
-    : Session.create({ title, model: modelInfo });
-  return session.id;
-}
-
-function resolveWorkerAgentName(target: Dispatch.Target): string | undefined {
-  return target.id ?? target.name;
-}
-
-function workerSpawnPayloadErrorMessage(error: z.ZodError<WorkerSpawnPayloadInput>): string {
-  if (error.issues.some((issue) => issue.code === "unrecognized_keys")) {
-    return "worker.spawn payload contains unsupported fields";
-  }
-
-  const fields = new Set(error.issues.flatMap((issue) => issue.path.map(String)));
-  if (fields.has("acceptanceCriteria")) {
-    return "worker.spawn requires at least one acceptance criterion";
-  }
-  if (fields.has("constraints")) {
-    return "worker.spawn constraints must be non-empty strings";
-  }
-  if (fields.has("text") || fields.has("prompt")) {
-    return WORKER_SPAWN_TEXT_OR_PROMPT_MESSAGE;
-  }
-
-  return "worker.spawn payload is invalid";
 }
 
 function parseWorkerCompletePayload(payload: unknown): WorkerCompletePayload {
@@ -120,37 +62,6 @@ function resolveCompletedWorkItemHash(
     throw new Error(`worker.complete could not resolve WorkItem for run ${workerRunId}`);
   }
   return workItem.hash;
-}
-
-function parseWorkerSpawnPayload(payload: unknown): ParsedWorkerSpawnPayload {
-  const parsed = WorkerSpawnPayload.safeParse(payload);
-  if (!parsed.success) {
-    const payloadRecord = payload && typeof payload === "object" ? payload : undefined;
-    if (!payloadRecord || !("acceptanceCriteria" in payloadRecord)) {
-      throw new Error("worker.spawn requires at least one acceptance criterion");
-    }
-    throw new Error(workerSpawnPayloadErrorMessage(parsed.error));
-  }
-
-  return parsed.data;
-}
-
-function buildRequest(
-  command: Dispatch.Command,
-  model: Model.Ref,
-  payload: ParsedWorkerSpawnPayload,
-): Execution.Request {
-  const sessionId = resolveSessionId(command, model);
-  return {
-    runId: crypto.randomUUID(),
-    sessionId,
-    mode: "direct",
-    prompt: payload.prompt,
-    model,
-    agentName: resolveWorkerAgentName(command.target),
-    workspaceRoot: command.workspaceRoot,
-    traceId: command.traceId,
-  };
 }
 
 export function createWorkerDispatchHandlers(
@@ -179,7 +90,7 @@ export function createWorkerDispatchHandlers(
       }
 
       const coordinator = requireCoordinator(options.coordinator);
-      const request = buildRequest(command, model, payload);
+      const request = buildWorkerSpawnRequest(command, model, payload);
       const workItemHash = await createWorkerSpawnWorkItem(
         command,
         request,
